requirements.txt
telethon
python-dotenv
ruff
mypy
